# internal/api/openapi/openapi_gin.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package openapi

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /shipments)
	CreateShipment(c *gin.Context)

	// (DELETE /shipments/{trackingNo})
	VoidShipment(c *gin.Context, trackingNo string)

	// (GET /shipments/{trackingNo})
	GetDocument(c *gin.Context, trackingNo string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// CreateShipment operation middleware
func (siw *ServerInterfaceWrapper) CreateShipment(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateShipment(c)
}

// VoidShipment operation middleware
func (siw *ServerInterfaceWrapper) VoidShipment(c *gin.Context) {

	var err error

	// ------------- Path parameter "trackingNo" -------------
	var trackingNo string

	err = runtime.BindStyledParameterWithOptions("simple", "trackingNo", c.Param("trackingNo"), &trackingNo, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter trackingNo: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.VoidShipment(c, trackingNo)
}

// GetDocument operation middleware
func (siw *ServerInterfaceWrapper) GetDocument(c *gin.Context) {

	var err error

	// ------------- Path parameter "trackingNo" -------------
	var trackingNo string

	err = runtime.BindStyledParameterWithOptions("simple", "trackingNo", c.Param("trackingNo"), &trackingNo, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter trackingNo: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetDocument(c, trackingNo)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/shipments", wrapper.CreateShipment)
	router.DELETE(options.BaseURL+"/shipments/:trackingNo", wrapper.VoidShipment)
	router.GET(options.BaseURL+"/shipments/:trackingNo", wrapper.GetDocument)
}