# internal/api/openapi/openapi_types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package openapi

// Defines values for WeightWeightUnit.
const (
	Kg WeightWeightUnit = "kg"
	Lb WeightWeightUnit = "lb"
)

// Defines values for CreateShipmentRequestShipmentPaymentInformationPaymentType.
const (
	Receiver   CreateShipmentRequestShipmentPaymentInformationPaymentType = "Receiver"
	Sender     CreateShipmentRequestShipmentPaymentInformationPaymentType = "Sender"
	ThirdParty CreateShipmentRequestShipmentPaymentInformationPaymentType = "ThirdParty"
)

// Defines values for CreateShipmentRequestShipmentPickupInformationPickupType.
const (
	DropOff      CreateShipmentRequestShipmentPickupInformationPickupType = "DropOff"
	PreScheduled CreateShipmentRequestShipmentPickupInformationPickupType = "PreScheduled"
)

// Defines values for CreateShipmentRequestPrinterType.
const (
	Regular CreateShipmentRequestPrinterType = "Regular"
	Thermal CreateShipmentRequestPrinterType = "Thermal"
)

// Defines values for DimensionDimensionUnit.
const (
	Cm DimensionDimensionUnit = "cm"
	In DimensionDimensionUnit = "in"
)

// Weight defines model for Weight.
type Weight struct {
	Value      int32            `json:"value"`
	WeightUnit WeightWeightUnit `json:"weightUnit"`
}

// WeightWeightUnit defines model for Weight.WeightUnit.
type WeightWeightUnit string

// Address defines model for Address.
type Address struct {
	Name         string  `json:"name" validate:"max=30"`
	Company      *string `json:"company,omitempty" validate:"max=20"`
	StreetNumber string  `json:"streetNumber" validate:"max=6"`
	StreetName   string  `json:"streetName" validate:"max=30"`
	City         string  `json:"city" validate:"max=30"`
	Province     string  `json:"province"`
	Country      string  `json:"country"`
	PostalCode   string  `json:"postalCode"`
	PhoneNumber  struct {
		CountryCode *string `json:"countryCode,omitempty"`
		AreaCode    *string `json:"areaCode,omitempty"`
		Phone       *string `json:"phone,omitempty"`
	} `json:"phoneNumber"`
}

// CreateShipmentRequest defines model for CreateShipmentRequest.
type CreateShipmentRequest struct {
	Shipment struct {
		SenderInformation struct {
			Address   Address `json:"address"`
			TaxNumber *string `json:"taxNumber,omitempty" validate:"max=11"`
		} `json:"senderInformation"`
		ReceiverInformation struct {
			Address   Address `json:"address"`
			TaxNumber *string `json:"taxNumber,omitempty" validate:"max=11"`
		} `json:"receiverInformation"`
		ShipmentDate       string `json:"shipmentDate"`
		PackageInformation struct {
			ServiceID         string `json:"serviceID"`
			Description       string `json:"description"`
			TotalWeight       Weight `json:"totalWeight"`
			TotalPieces       int32  `json:"totalPieces"`
			PiecesInformation *struct {
				Pieces []Piece `json:"pieces" xml:"Piece"`
			} `json:"piecesInformation,omitempty"`
		} `json:"packageInformation"`
		PaymentInformation struct {
			PaymentType             *CreateShipmentRequestShipmentPaymentInformationPaymentType `json:"paymentType,omitempty"`
			RegisteredAccountNumber *string                                                     `json:"registeredAccountNumber,omitempty"`
			BillingAccountNumber    *string                                                     `json:"billingAccountNumber,omitempty"`
		} `json:"paymentInformation"`
		PickupInformation struct {
			PickupType *CreateShipmentRequestShipmentPickupInformationPickupType `json:"pickupType,omitempty"`
		} `json:"pickupInformation"`
		TrackingReferenceInformation *struct {
			Reference1 *string `json:"reference1,omitempty"`
			Reference2 *string `json:"reference2,omitempty"`
			Reference3 *string `json:"reference3,omitempty"`
			Reference4 *string `json:"reference4,omitempty"`
		} `json:"trackingReferenceInformation,omitempty"`
	} `json:"shipment"`
	PrinterType CreateShipmentRequestPrinterType `json:"printerType"`
}

// CreateShipmentRequestShipmentPaymentInformationPaymentType defines model for CreateShipmentRequest.Shipment.PaymentInformation.PaymentType.
type CreateShipmentRequestShipmentPaymentInformationPaymentType string

// CreateShipmentRequestShipmentPickupInformationPickupType defines model for CreateShipmentRequest.Shipment.PickupInformation.PickupType.
type CreateShipmentRequestShipmentPickupInformationPickupType string

// CreateShipmentRequestPrinterType defines model for CreateShipmentRequest.PrinterType.
type CreateShipmentRequestPrinterType string

// Dimension defines model for Dimension.
type Dimension struct {
	Value         int32                   `json:"value"`
	DimensionUnit *DimensionDimensionUnit `json:"dimensionUnit,omitempty"`
}

// DimensionDimensionUnit defines model for Dimension.DimensionUnit.
type DimensionDimensionUnit string

// Error defines model for Error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// Piece defines model for Piece.
type Piece struct {
	Weight Weight    `json:"weight"`
	Height Dimension `json:"height"`
	Length Dimension `json:"length"`
	Width  Dimension `json:"width"`
}

// CreateShipmentJSONRequestBody defines body for CreateShipment for application/json ContentType.
type CreateShipmentJSONRequestBody = CreateShipmentRequest